+++ Lib/Sources/UIUnitTest/Element/Element.swift
@@ -17,25 +17,6 @@ public class Element: ElementTypeQueryProvider, @unchecked Sendable {
         }
     }
     
-    public func exists() async throws ->  Bool {
-        let existsRequestData = ElementRequest(serverId: serverId)
-        let existsResponse: ExistsResponse = try await callServer(path: "exists", request: existsRequestData)
-        return existsResponse.exists
-    }
-    
-    /** Waits the specified amount of time for the element's exist property to be true and returns false if the timeout expires without the element coming into existence. */
-    public func waitForExistence(timeout: TimeInterval) async throws -> Bool {
-        let activateRequestData = WaitForExistenceRequest(serverId: serverId, timeout: timeout)
-        let response: WaitForExistenceResponse = try await callServer(path: "waitForExistence", request: activateRequestData)
-        return response.elementExists
-    }
-    
-    public func waitForNonExistence(timeout: TimeInterval) async throws -> Bool {
-        let activateRequestData = WaitForExistenceRequest(serverId: serverId, timeout: timeout)
-        let response: WaitForExistenceResponse = try await callServer(path: "waitForNonExistence", request: activateRequestData)
-        return !response.elementExists
-    }
-    
     /** Whether or not a hit point can be computed for the element for the purpose of synthesizing events. */
     public func isHittable() async throws -> Bool  {
         let existsRequestData = ElementRequest(serverId: serverId)
