+++ Lib/Sources/UIUnitTest/Query/Query+QueryType.swift
